msm: kgsl: Avoid inline submission if slumber is in progress
Consider a scenario where usermode is submitting commands while slumber
is in progress. During slumber sequence, device mutex is held by kgsl and
it polls for CX gdsc to go down. It is possible that SMMU can keep a vote
on CX due to ongoing iommu operations and it may take longer for CX gdsc
to go down.

In another thread, say usermode is submitting commands, currently these
commands are submitted inline to the GPU for gmu based targets. However,
while submitting commands to GPU, device mutex is needed. Usermode thread
will be blocked on device mutex as it is held by slumber thread. This can
result in UI freeze or janks.

To resolve this, avoid inline submission whenever slumber is in progress.
Add relevant code to properly update the slumber entry and exit flag.

Change-Id: I1b103891ad4b386e230b260fe5dbf6db20470ca9
Signed-off-by: Deepak Kumar <quic_dkumar@quicinc.com>
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>
Signed-off-by: Forenche <prahul2003@gmail.com>